Uses metadata.name if exists as secret name
... instead of using the Puppet resource's title only for the name. This
is useful if you want to have the same secret (same name) in multiple
namespaces. Puppet will not allow you to have multiple resources with
the same name and currently the resource doesn't respect the name in the
metadata spec. This change checks if metadata.name exists and uses that
for the secret name. If it doesn't exist, it defaults to using the
Puppet resource's name. The drawback is that the user needs to pass the
puppet resource name into the secret metadata so that Puppet can check
if it's managing it or not (because the secret name doesn't match the
puppet resource anymore). This won't affect the default behavior.