Changes name param to secretname
The name param is a special Puppet param that is a synonym of the title.
This prevents us from calling the secret something different than the
title. This is useful when you want to have the same secret name in
multiple namespaces and must therefore create multiple
kubernetes_resources with different names to avoid duplicate resource
declaration errors.